@@ -1,6 +1,22 @@
 ﻿namespace Seatbelt.Commands
 {
-    public interface CommandDTOBase
+    public class CommandDTOBase
     {
+        private string CommandVersion { get; set; }
+
+        protected CommandDTOBase()
+        {
+            CommandVersion = "";
+        }
+
+        public void SetCommandVersion(string commandVersion)
+        {
+            CommandVersion = commandVersion;
+        }
+        
+        public string GetCommandVersion()
+        {
+            return CommandVersion;
+        }
     }
 }
