@@ -22,50 +22,46 @@ namespace Seatbelt.Commands
 
         private IEnumerable<WindowsSearchIndexDTO> SearchWindowsIndex(string searchPath = @"C:\Users\", string criteria = "password")
         {
-            var Query = @"SELECT System.ItemPathDisplay,System.FileOwner,System.Size,System.DateCreated,System.DateAccessed,System.Search.Autosummary FROM SystemIndex WHERE Contains(*, '""*{0}*""') AND SCOPE = '{1}' AND (System.FileExtension = '.txt' OR System.FileExtension = '.doc' OR System.FileExtension = '.docx' OR System.FileExtension = '.ppt' OR System.FileExtension = '.pptx' OR System.FileExtension = '.xls' OR System.FileExtension = '.xlsx' OR System.FileExtension = '.ps1' OR System.FileExtension = '.vbs' OR System.FileExtension = '.config' OR System.FileExtension = '.ini')";
-
+            var format = @"SELECT System.ItemPathDisplay,System.FileOwner,System.Size,System.DateCreated,System.DateAccessed,System.Search.Autosummary FROM SystemIndex WHERE Contains(*, '""*{0}*""') AND SCOPE = '{1}' AND (System.FileExtension = '.txt' OR System.FileExtension = '.doc' OR System.FileExtension = '.docx' OR System.FileExtension = '.ppt' OR System.FileExtension = '.pptx' OR System.FileExtension = '.xls' OR System.FileExtension = '.xlsx' OR System.FileExtension = '.ps1' OR System.FileExtension = '.vbs' OR System.FileExtension = '.config' OR System.FileExtension = '.ini')";
             var connectionString = "Provider=Search.CollatorDSO;Extended Properties=\"Application=Windows\"";
-            using (var connection = new OleDbConnection(connectionString))
-            {
-                var query = string.Format(Query, criteria, searchPath);
+            
+            using var connection = new OleDbConnection(connectionString);
+            var query = string.Format(format, criteria, searchPath);
 #pragma warning disable CA2100
-                var command = new OleDbCommand(query, connection);
+            var command = new OleDbCommand(query, connection);
 #pragma warning restore CA2100
-                connection.Open();
+            connection.Open();
 
-                var result = new List<string>();
-
-                OleDbDataReader reader = null;
-                try
-                {
-                    reader = command.ExecuteReader();
-                }
-                catch
-                {
-                    WriteError("Unable to query the Search Indexer, Search Index is likely not running.");
-                    yield break;
-                }
-
-                while (reader.Read())
-                {
-                    var AutoSummary = "";
-                    var FileOwner = "";
-                    try { AutoSummary = reader.GetString(5); } catch { }
-                    try { FileOwner = reader.GetString(1); } catch { }
-
-                    yield return new WindowsSearchIndexDTO()
-                    {
-                        Path = reader.GetString(0),
-                        FileOwner = FileOwner,
-                        Size = Decimal.ToUInt64((Decimal)reader.GetValue(2)),
-                        DateCreated = reader.GetDateTime(3),
-                        DateAccessed = reader.GetDateTime(4),
-                        AutoSummary = AutoSummary
-                    };
-                }
-
-                connection.Close();
+            OleDbDataReader reader;
+            try
+            {
+                reader = command.ExecuteReader();
             }
+            catch
+            {
+                WriteError("Unable to query the Search Indexer, Search Index is likely not running.");
+                yield break;
+            }
+
+            while (reader.Read())
+            {
+                var AutoSummary = "";
+                var FileOwner = "";
+                try { AutoSummary = reader.GetString(5); } catch { }
+                try { FileOwner = reader.GetString(1); } catch { }
+
+                yield return new WindowsSearchIndexDTO()
+                {
+                    Path = reader.GetString(0),
+                    FileOwner = FileOwner,
+                    Size = Decimal.ToUInt64((Decimal)reader.GetValue(2)),
+                    DateCreated = reader.GetDateTime(3),
+                    DateAccessed = reader.GetDateTime(4),
+                    AutoSummary = AutoSummary
+                };
+            }
+
+            connection.Close();
         }
 
 
