@@ -12,7 +12,7 @@ namespace Seatbelt.Commands.Windows
     internal class NamedPipesCommand : CommandBase
     {
         public override string Command => "NamedPipes";
-        public override string Description => "Named pipe names and any readable ACL information.";
+        public override string Description => "Named pipe names, any readable ACL information and associated process information.";
         public override CommandGroup[] Group => new[] {CommandGroup.System};
 
         public override bool SupportRemote => false; // almost certainly not possible
@@ -42,6 +42,62 @@ namespace Seatbelt.Commands.Windows
             {
                 FileSecurity security;
                 var sddl = "";
+                int iProcessId = 0;
+                string svProcessName = "";
+                string svProcessPath = "";
+                System.IntPtr hPipe = System.IntPtr.Zero;
+                bool bvRet = false;
+
+                // Try to identify ProcessID and ProcessName
+                try
+                {
+                    //Get a handle to the pipe
+                    hPipe = CreateFile(
+                        System.String.Format("\\\\.\\pipe\\{0}", namedPipe), // The name of the file or device to be created or opened.
+                        FileAccess.Read, // The requested access to the file or device.
+                        FileShare.None, // The requested sharing mode of the file or device.
+                        System.IntPtr.Zero, // Optional. A pointer to a SECURITY_ATTRIBUTES structure.
+                        FileMode.Open, // An action to take on a file or device that exists or does not exist.
+                        FileAttributes.Normal, // The file or device attributes and flags.
+                        System.IntPtr.Zero); // Optional. A valid handle to a template file with the GENERIC_READ access right.
+
+
+                    if (hPipe.ToInt64() != -1) //verify CreateFile did not return "INVALID_HANDLE_VALUE"
+                    { 
+
+                        //Retrieve the ProcessID registered for the pipe.
+                        bvRet = GetNamedPipeServerProcessId(
+                            hPipe, // A handle to an instance of a named pipe.
+                            out iProcessId); // The process identifier.
+
+                        //If GetNamedPipeServerProcessId was successful, get the process name for the returned ProcessID
+                        if (bvRet)
+                        {
+                            var svProcess = System.Diagnostics.Process.GetProcessById(iProcessId);
+                            svProcessName = svProcess.ProcessName;
+                            svProcessPath = svProcess.MainModule.FileName;
+                        }
+                        else
+                        {
+                            //GetNamedPipeServerProcessId was unsuccessful
+                            svProcessName = "Unk";
+                        }
+
+                        //Close the pipe handle
+                        CloseHandle(hPipe);
+                    }
+                    else
+                    {
+                        //CreateFile returned "INVALID_HANDLE_VALUE" or 0xffffffff.
+                        svProcessName = "Unk";
+                    }
+                }
+                catch
+                {
+                    //Catch the exception. ProcessName is set to Unk.
+                    svProcessName = "Unk";
+                }
+
                 try
                 {
                     security = File.GetAccessControl(System.String.Format("\\\\.\\pipe\\{0}", namedPipe));
@@ -52,24 +108,16 @@ namespace Seatbelt.Commands.Windows
                     sddl = "ERROR";
                 }
 
-                if (!System.String.IsNullOrEmpty(sddl) && !sddl.Equals("ERROR"))
+                yield return new NamedPipesDTO()
                 {
-                    yield return new NamedPipesDTO()
-                    {
-                        Name = namedPipe,
-                        Sddl = sddl
-                        //SecurityDescriptor = new RawSecurityDescriptor(sddl)
-                    };
-                }
-                else
-                {
-                    yield return new NamedPipesDTO()
-                    {
-                        Name = namedPipe,
-                        Sddl = sddl
-                        //SecurityDescriptor = null
-                    };
-                }
+                    Name = namedPipe,
+                    Sddl = sddl,
+                    //SecurityDescriptor = null
+
+                    ServerProcessName = svProcessName,
+                    ServerProcessPID = iProcessId,
+                    ServerProcessPath = svProcessPath
+                };
             }
         }
     }
@@ -79,6 +127,12 @@ namespace Seatbelt.Commands.Windows
         public string Name { get; set; }
 
         public string Sddl { get; set; }
+
+        public string ServerProcessName { get; set; }
+
+        public int ServerProcessPID { get; set; }
+
+        public string ServerProcessPath { get; set; }
 
         // public RawSecurityDescriptor SecurityDescriptor { get; set; }
     }
@@ -94,7 +148,16 @@ namespace Seatbelt.Commands.Windows
         {
             var dto = (NamedPipesDTO)result;
 
-            WriteLine("  \\\\.\\pipe\\{0}", dto.Name);
+            WriteLine("\n{0}", dto.Name);
+            WriteLine("    Server Process Id   : {0}", dto.ServerProcessPID.ToString());
+            if (!string.IsNullOrEmpty(dto.ServerProcessPath))
+            {
+                WriteLine("    Server Process Name : {0}", dto.ServerProcessName);
+            }
+            if (!string.IsNullOrEmpty(dto.ServerProcessPath))
+            {
+                WriteLine("    Server Process Path : {0}", dto.ServerProcessPath);
+            }
             if (!dto.Sddl.Equals("ERROR"))
             {
                 //WriteLine("    Owner   : {0}", dto.SecurityDescriptor.Owner);
@@ -103,7 +166,7 @@ namespace Seatbelt.Commands.Windows
                 //    WriteLine("        {0} :", rule.SecurityIdentifier);
                 //    WriteLine("              {0} : {1}", rule.AceType, (GenericAceMask)rule.AccessMask);
                 //}
-                WriteLine("    SDDL         : {0}", dto.Sddl);
+                WriteLine("    Pipe SDDL           : {0}", dto.Sddl);
             }
         }
     }
