@@ -88,7 +88,10 @@ namespace Seatbelt.Output.Sinks
         public void WriteError(string message) => WriteOutput(new ErrorDTO(message));
 
         public void WriteHost(string message) => WriteOutput(new HostDTO(message));
-
+        public string GetOutput()
+        {
+            return "";
+        }
         public void Dispose()
         {
             _textWriter.Dispose();
